fix(cc_resize): Don't warn if zpool command not found (#4969)
Not resizing the rootfs should not cause warnings in containers.
This is expected, and how cloud-init currently behaves. Similarly,
if the image doesn't include the tools to resize the rootfs in a
container, cloud-init should similarly not raise warnings.

Other changes:
- relocate single-use helper function from cc_resize to util.py
- add warning on non-containers when no device is found.

LP: #2055219
